🎨 Update interfaces and utils listPost to NewsWrapper
