--- src/components/listPosts/interfaces.ts
@@ -1,4 +1,0 @@
-export interface optionsInterface {
-    label: string
-    value: number
-}
