--- src/components/listPosts/listPosts.module.sass
@@ -1,20 +1,0 @@
-@import 'styles/vars'
-
-.tabsWrapper
-  display: flex
-  align-items: center
-  justify-content: center
-  margin-bottom: $spacing-m * 3.938
-
-.tab
-  border: 1px solid $gray-light
-  width: $spacing-m * 6.125
-  padding: $spacing-xxs
-  display: flex
-  justify-content: center
-  color: $gray
-  cursor: pointer
-
-.selected
-  color: $blue
-  border-color: $blue
