--- src/components/listPosts/utils.ts
@@ -1,10 +1,0 @@
-export const options = [
-    {
-        label: 'All',
-        value: 0
-    },
-    {
-        label: 'My faves',
-        value: 1
-    }
-]
